Crate webserver

Server for Maximise website.

Routes

/users

/create - creates new user

/list - lists all usernames

/login - logs in user with username and password, returning a session ID

/delete - deletes user

/logout - logs out user from all sessions

/change_password - replaces a user's password

/get_data - returns data stored for a user

/update_data - replaces currently stored data for a user with new data

/codes

/get_next - gets next available code and marks it as sold

For more information see the database module.

Usage of /users

Generally the website will be using the /users section of the API. Every route in the /users section apart from /get_data and update_data is for high level access. High level access involves authenticating users, creating, deleting and modifying critical user information, such as passwords.

The /get_data and /update_data routes are for low level access. This allows easy access to other user data, such as information stored by the website itself. This includes email addresses.

In order to read or modify any user data, the user must first be authenticated through /login. If successful, a session ID will be provided to the client as an easy means of access to the user data for the current client session. In this case, session ID is used like a temporary password, so username must still be provided every time it is used. The reasoning behind this is that the bcrypt hashing algorithm is designed to take a long time to hash and verify passwords, meaning that if the username and password is provided for every single access to the server, the user experience will be incredibly slow, as every action they make would take upwards of 10 seconds. The session ID is instead used as a proof that the user has been authenticated on that client recently, and verifying the session ID involves no hashing, meaning that it is much faster. Obviously the sesssions are frequently reset and therefore sessions will expire eventually. Every time the user logs out of the website, they should also be logged out on the server-side, using /logout so that their sessions can be deleted.

For more information see the database module.

Usage of /codes/get_next

This is not designed for use on the website. This is designed as an easy way to both generate codes and let the server know when codes are generated. We could either put the codes in the box, or when selling the box, we could simply get the email address of the customer, generate a code and email the code to them.

The codes are not necessarily unique to each user, but at any one time, the codes available to use are all different. This means that codes should not be used as a unique user ID.

General information

At any given time, it should be assumed that the server holds the most up-to-date version of the data, even if the data in the database does not match with the data in the server. This is because the database is updated a minimum of once a day, which means that it can often hold out-of-date information. The database should never be directly accessed anyway, because that is what this server is for.

The database cannot be directly accessed, because the database access rules only allow for the server to access it.

The server will be restarted once a day, meaning that it will go offline very shortly once a day, at some off-peak time. Furthermore, the server will go to sleep after one hour of no requests - when the server is asleep, it may take a few more seconds than normal for it to respond to the first request. The best way to counteract this is to ping the server with /users/list when a user clicks the login/signup button. This will not only wake up the server, but the client can then immediately respond to the username entered, warning if the username is taken or doesn't exist as appropriate.

Modules

database
Handles all requests.

Functions

authenticate

deauthenticate

main