Module webserver::database

Handles all requests.

This module contains handlers for all the requests that the server handles. Two types of requests can be used:

GET: used for simple requests where no private data needs to be passed to the server.

POST: used for more complex requests where data can be passed to the server in JSON form.

Structs

State
This is a container struct for the current state that the server is using.

Functions

change_user_password
Accessed via POST /users/change_password: replaces a user's password in the database, and immediately logs them out.

create_user
Accessed via POST /users/create: creates a user in the database.

delete_user
Accessed via POST /users/delete: deletes a user from the database.

generate_code
Helper function to generate a random code. Not part of API.

get_next_code
Accessed via GET /codes/get_next: responds with a three-word code that can be supplied to the box and adds this code to the database.

get_user_data
Accessed via POST /users/get_data: returns a JSON object containing data stored for a user.

list_users
Accessed via GET /users/list: responds with a list of users currently in the database.

login_user
Accessed via POST /users/login: responds with a session ID for the client to use for further database interactions.

logout_user
Accessed via POST /users/logout: logs out a user, removing all their session IDs.

update_user_data
Accessed via POST /users/update_data: replaces the currently stored data for a user with the JSON object provided.