Function webserver::database::get_next_code

pub async fn get_next_code(state: Data<State>) -> String

Accessed via GET /codes/get_next: responds with a three-word code that can be supplied to the box and adds this code to the database.

Request fields: none

Response fields: response (array(string)), errors (array(string))

Example response

The following response contains no errors, meaning that the request was successful.

{
    "response": ["big-green-zebra"],
    "errors": []
}

Errors

Internal errors are caused by errors in the server. External errors are caused by errors in the request. External errors are much more likely to happen.

Internal

State must in JSON format - very unlikely that this happens.