Function webserver::database::get_user_data

pub async fn get_user_data(request: String, state: Data<State>) -> String

Accessed via POST /users/get_data: returns a JSON object containing data stored for a user.

Request fields: username (string), sessionID (string)

Response fields: response (JSON), errors (array(string))

Example request

{
    "username": "my_name",
    "sessionID": "123abc456def"
}

Example responses

Error response

The following response contains an error, meaning that the request failed.

{
    "response": ""
    "errors": ["External: Session not found"]
}

Ok response

The following response has no errors, meaning that the request was successful.

{
    "response": "{"email": "someone@somewhere.com"}",
    "errors": []
}

Errors

Internal errors are caused by errors in the server. External errors are caused by errors in the request. External errors are much more likely to happen.

Internal

State must be in JSON format - very unlikely that this happens.

External:

Request must be in JSON format - the server couldn't convert the request into JSON.

Request must contain username and session ID - the server couldn't find the required username and session ID fields in the request JSON.

Username must be string - the server couldn't convert the username in the request JSON to a string.

Session ID must be string - the server couldn't convert the session ID in the request JSON to a string.

Session not found - the session ID provded in the request JSON was not found in the database. This is probably because the session has expired.

Incorrect username for session ID; logging out - the session ID provided in the request JSON didn't match with the username provided in the JSON. This would usually only happen because of a malicious attack on the data, so all users are immediately logged out.

No data - the user has no stored data.